make skills version in-dependant
