+++ docs/wiki/factions/hunter/skills.mdx
@@ -4,7 +4,8 @@ sidebar_label: Skills
 title: Hunter Skills
 ---
 
-import { Skill, Action, SkillBranch, UnlockAction} from "@site/src/components/Skills";
+import { SkillBranch, UnlockAction} from "@site/src/components/Skills";
+import { Skill, Action } from "../_skills";
 
 Main article [here](../skills)
