+++ docs/wiki/factions/vampire/skills.mdx
@@ -4,7 +4,8 @@ title: Vampire Skills
 sidebar_label: Skills
 ---
 
-import { Skill, Action, UnlockAction, SkillBranch, UnlocksVision, UpgradesSkill } from "@site/src/components/Skills";
+import { UnlockAction, SkillBranch, UnlocksVision, UpgradesSkill } from "@site/src/components/Skills";
+import { Skill, Action } from "../_skills";
 
 Main article [here](../skills)
