+++ src/components/Skills.tsx
@@ -1,11 +1,5 @@
 import React from "react";
 
-export const baseUrl = 'https://raw.githubusercontent.com/TeamLapen/Vampirism/'
-export const baseCommit = '4bf2c73fb860a23de225edbae9c0b1c1ead3dd1a'
-export const skillPath = '/src/main/resources/assets/vampirism/textures/skills/'
-export const actionPath = '/src/main/resources/assets/vampirism/textures/actions/'
-
-
 export const SkillImg = ({src, alt}) => (
     <img src={src} alt={alt} style={{height: 70, imageRendering: "pixelated"}}/>
 );
@@ -52,18 +46,6 @@ export const SkillOrAction = ({children, title, iconPath}) => (
         </span>
     </div>
 );
-
-export function Skill({children, title, icon, commit=baseCommit}) {
-    return (<SkillOrAction title={title} iconPath={baseUrl.concat(commit).concat(skillPath).concat(icon)}>
-        {children}
-    </SkillOrAction>);
-}
-
-export function Action({children, title, icon, commit=baseCommit}) {
-    return(<SkillOrAction title={title} iconPath={baseUrl.concat(commit).concat(actionPath).concat(icon)}>
-        {children}
-    </SkillOrAction>);
-}
 
 export const SkillBranch = ({children, title}) =>(
     <div>
